Added combinePath
