+++ libs/dh/core/routing/src/index.ts
@@ -1,72 +1,1 @@
-/**
- * @license
- * Copyright 2020 Energinet DataHub A/S
- *
- * Licensed under the Apache License, Version 2.0 (the "License2");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-const marketParticipantSubPaths = {
-  actorsPath: 'actors',
-  organizationsPath: 'organizations',
-  marketRolesPath: 'market-roles',
-} as const;
-
-const eSettSubPaths = {
-  outgoingMessagesPath: 'outgoing-messages',
-  meteringGridareaImbalancePath: 'metering-gridarea-imbalance',
-  imbalanceResponsiblePartiesPath: 'balance-responsible',
-} as const;
-
-const adminSubPaths = {
-  user: 'user',
-  roles: 'roles',
-  permissions: 'permissions',
-} as const;
-
-const wholesaleSubPaths = {
-  requestCalculation: 'request-calculation',
-  calculations: 'calculations',
-  settlementReports: 'settlement-reports',
-} as const;
-
-const basePaths = {
-  marketParticipantBasePath: 'market-participant',
-  messageArchiveBasePath: 'message-archive',
-  esettBasePath: 'esett',
-  admin: 'admin',
-  imbalancPrices: 'imbalance-prices',
-  gridAreas: 'grid-areas',
-  wholesale: 'wholesale',
-  login: 'login',
-} as const;
-
-export type MarketParticipantSubPaths =
-  (typeof marketParticipantSubPaths)[keyof typeof marketParticipantSubPaths];
-
-export type BasePaths = (typeof basePaths)[keyof typeof basePaths];
-
-export type ESettSubPaths = (typeof eSettSubPaths)[keyof typeof eSettSubPaths];
-
-export type WholesaleSubPaths = (typeof wholesaleSubPaths)[keyof typeof wholesaleSubPaths];
-
-export type AdminSubPaths = (typeof adminSubPaths)[keyof typeof adminSubPaths];
-
-export const getPath = <
-  T extends
-    | BasePaths
-    | MarketParticipantSubPaths
-    | ESettSubPaths
-    | WholesaleSubPaths
-    | AdminSubPaths,
->(
-  route: T
-) => route;
+export * from './dhPaths';
