+++ libs/dh/core/shell/src/lib/dh-primary-navigation.component.ts
@@ -20,7 +20,7 @@ import { TranslocoDirective } from '@ngneat/transloco';
 import { WattNavListComponent, WattNavListItemComponent } from '@energinet-datahub/watt/shell';
 import { DhFeatureFlagDirective } from '@energinet-datahub/dh/shared/feature-flags';
 import { DhPermissionRequiredDirective } from '@energinet-datahub/dh/shared/feature-authorization';
-import { WholesaleSubPaths, getPath, type BasePaths } from '@energinet-datahub/dh/core/routing';
+import { WholesaleSubPaths, combinePaths, BasePaths } from '@energinet-datahub/dh/core/routing';
 
 @Component({
   changeDetection: ChangeDetectionStrategy.OnPush,
@@ -47,5 +47,5 @@ export class DhPrimaryNavigationComponent {
   getLink(route: BasePaths) {
     return `/${route}`;
   }
-  getWholesaleLink = (path: WholesaleSubPaths) => `${this.getLink('wholesale')}/${getPath(path)}`;
+  getWholesaleLink = (path: WholesaleSubPaths) => combinePaths('wholesale', path);
 }
